Revert "Shush another compiler warning.", replace it with inttypes.h macros.
This warning is correct -- you're feeding a 64 bit value to a format
specifier that can mean either 64 or 128 bits, depending on platform.

This reverts commit f142bfad6f59f60829fc1980bb4178f51e6a651a.

+++ crawl-ref/source/database.cc
@@ -245,7 +245,7 @@ bool TextDB::_needs_update() const
         if (mtime)
             no_files = false;
         char buf[20];
-        snprintf(buf, sizeof(buf), ":%ld", mtime);
+        snprintf(buf, sizeof(buf), ":%"PRId64, (int64_t)mtime);
         ts += buf;
     }
 
@@ -296,7 +296,7 @@ void TextDB::_regenerate_db()
         full_input_path = datafile_path(full_input_path, !_parent);
         char buf[20];
         time_t mtime = file_modtime(full_input_path);
-        snprintf(buf, sizeof(buf), ":%ld", mtime);
+        snprintf(buf, sizeof(buf), ":%"PRId64, (int64_t)mtime);
         ts += buf;
         if (mtime || !_parent) // english is mandatory
             _store_text_db(full_input_path, _db);
